+++ common/src/main/java/io/github/flemmli97/tenshilib/patreon/client/PatreonClientPlatform.java
@@ -1,9 +1,9 @@
 package io.github.flemmli97.tenshilib.patreon.client;
 
+import io.github.flemmli97.tenshilib.common.network.Packet;
 import io.github.flemmli97.tenshilib.mixinhelper.ScreenWidgetAdder;
 import io.github.flemmli97.tenshilib.patreon.PatreonPlatform;
 import io.github.flemmli97.tenshilib.patreon.RenderLocation;
-import io.github.flemmli97.tenshilib.patreon.pkts.C2SEffectUpdatePkt;
 import io.github.flemmli97.tenshilib.platform.InitUtil;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.SkinCustomizationScreen;
@@ -33,5 +33,5 @@ public interface PatreonClientPlatform {
 
     void setup();
 
-    void sendToServer(C2SEffectUpdatePkt pkt);
+    void sendToServer(Packet pkt);
 }