+++ Changelog.md
@@ -1,6 +1,7 @@
 TenshiLib 1.6.15
 ================
 - Add itemstack and entity sensitive version to some stuff
+- Add a color slider and alpha support for patreon effects
 
 TenshiLib 1.6.14
 ================